Refactor Kotlin Paging DSL
