+++ src/main/kotlin/org/mybatis/dynamic/sql/util/kotlin/KotlinPagingDSL.kt
@@ -1,0 +1,36 @@
+/*
+ *    Copyright 2016-2024 the original author or authors.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+package org.mybatis.dynamic.sql.util.kotlin
+
+interface KotlinPagingDSL {
+    fun limit(limit: Long) {
+        limitWhenPresent(limit)
+    }
+
+    fun limitWhenPresent(limit: Long?)
+
+    fun offset(offset: Long) {
+        offsetWhenPresent(offset)
+    }
+
+    fun offsetWhenPresent(offset: Long?)
+
+    fun fetchFirst(fetchFirstRows: Long) {
+        fetchFirstWhenPresent(fetchFirstRows)
+    }
+
+    fun fetchFirstWhenPresent(fetchFirstRows: Long?)
+}
